chore: 增加 Core 工程
