+++ src/BootstrapBlazor.Core/BootstrapBlazor.Core.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <Version>8.0.0</Version>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Using Include="BootstrapBlazor.Components" />
+  </ItemGroup>
+
+</Project>
