+++ src/BootstrapBlazor/Extensions/EnumExtensions.cs
@@ -13,31 +13,6 @@ namespace BootstrapBlazor.Components;
 public static class EnumExtensions
 {
     /// <summary>
-    /// 获取 DescriptionAttribute 标签方法
-    /// </summary>
-    /// <param name="val"></param>
-    /// <returns></returns>
-    public static string ToDescriptionString<TEnum>(this TEnum val) where TEnum : Enum => typeof(TEnum).ToDescriptionString(val.ToString());
-
-    /// <summary>
-    /// 通过字段名称获取 DescriptionAttribute 标签值
-    /// </summary>
-    /// <param name="type"></param>
-    /// <param name="fieldName"></param>
-    /// <returns></returns>
-    public static string ToDescriptionString(this Type? type, string? fieldName)
-    {
-        var ret = string.Empty;
-        if (type != null && !string.IsNullOrEmpty(fieldName))
-        {
-            var t = Nullable.GetUnderlyingType(type) ?? type;
-            var attributes = t.GetField(fieldName)?.GetCustomAttribute<DescriptionAttribute>(true);
-            ret = attributes?.Description ?? fieldName;
-        }
-        return ret;
-    }
-
-    /// <summary>
     /// 通过字段名称获取 DisplayAttribute/DescriptionAttribute 标签值
     /// </summary>
     /// <typeparam name="TEnum"></typeparam>
