--- src/Extensions/Components/BootstrapBlazor.Core/BootstrapBlazor.Core.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Razor">
-
-  <PropertyGroup>
-    <Version>8.0.0-beta01</Version>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <PackageTags>Bootstrap Blazor WebAssembly wasm UI Components</PackageTags>
-    <Description>Bootstrap UI components core libarayrary</Description>
-  </PropertyGroup>
-
-</Project>
